+++ api_app/connectors_manager/connectors/email_sender.py
@@ -1,27 +1,23 @@
 from django.core.mail import EmailMessage
 
 from api_app.connectors_manager.classes import Connector
+from intel_owl.settings import DEFAULT_FROM_EMAIL
 from tests.mock_utils import if_mock_connections, patch
 
 
 class EmailSender(Connector):
     sender: str
-
-    @property
-    def subject(self) -> str:
-        return f"Take down domain {self._job.parent_job.parent_job.observable_name}"
-
-    @property
-    def body(self) -> str:
-        return (
-            f"Domain {self._job.parent_job.parent_job.observable_name} "
-            f"has been reported as malicious. We request you to take it down."
-        )
+    subject: str
+    body: str
 
     def run(self) -> dict:
+        if self.sender:
+            sender = self.sender
+        else:
+            sender = DEFAULT_FROM_EMAIL
         base_eml = EmailMessage(
             subject=self.subject,
-            from_email=self.sender,
+            from_email=sender,
             to=[self._job.observable_name],
             body=self.body,
         )
